refactor(ms2/deployments): moved element status to separate file
